refactor: move version module to services and update imports
