+++ src/hivebox/services/__init__.py
@@ -1,0 +1,1 @@
+"""Services module for hivebox."""
